Queue `ping` tasks with a current expiry timestamp.
We used to queue them with a fake expiry of `1` but this could
mess up code that expected a realistic value.  It's not clear
why we need to set a value at all (since most tasks are queued
with no expiry and gain one only when leased) but setting it to
the current timestamp instead seems reasonable.  The tasks will
still be immediately eligible for leasing.